+++ app/budget/views.py
@@ -6,9 +6,13 @@ from rest_framework import viewsets
 from home.globals import functions
 from django_filters import rest_framework as filters
 from .filter import BudgetFilter
-from home.views import AuthenticateView
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
+from rest_framework_jwt.authentication import JSONWebTokenAuthentication
+from home.globals import const
+from rest_framework.response import Response
 
-class BudgetViewSet(viewsets.ModelViewSet, AuthenticateView):
+
+class BudgetViewSet(viewsets.ModelViewSet):
     serializer_class = BudgetSerializer
     lookup_field = 'user'
     filter_backends = (filters.DjangoFilterBackend,)
@@ -26,4 +30,15 @@ class BudgetViewSet(viewsets.ModelViewSet, AuthenticateView):
         serializer.save(user=self.request.user)
 
 
+@api_view(['GET'])
+@authentication_classes((JSONWebTokenAuthentication, ))
+def budget_types(request):
+    """
+    get:
+        get BudgetTypeData
+    """
+    if request.method == 'GET':
+        budget_type = const.budget_type
+
+        return Response(budget_type)
