+++ app/transaction/views.py
@@ -4,7 +4,10 @@ from .serializer import TransactionSerializer
 from rest_framework import viewsets
 from django_filters import rest_framework as filters
 from .filter import TransactionFilter
-
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
+from rest_framework_jwt.authentication import JSONWebTokenAuthentication
+from home.globals import const
+from rest_framework.response import Response
 
 
 # Create your views here.
@@ -22,3 +25,14 @@ class TransactionViewSet(viewsets.ModelViewSet):
     def perform_create(self, serializer):
         serializer.save(user=self.request.user)
 
+
+@api_view(['GET'])
+@authentication_classes((JSONWebTokenAuthentication, ))
+def transaction_types(request):
+    """
+    get:
+        get TransactionTypeData
+    """
+    if request.method == 'GET':
+        transaction_type = const.transaction_type
+        return Response(transaction_type)
